GLWidget fixed

+++ Block.cpp
@@ -3,7 +3,7 @@
 Block::Block(){}
 
 
-Block::Block(Point3 p_min, double size_x,double size_y,double size_z): size_x(size_x), 
+Block::Block(Point3 p_min, float size_x,float size_y,float size_z): size_x(size_x), 
 size_y(size_y), 
 size_z(size_z) {}
